- Added Clock class to manage time-based actions - Updated Application.cpp to initialize the Clock instance - Enhanced input state access, events tracking in ImGuiLayer.cpp - Added more functionalities in DebugLayer for button presses - Minor code cleanups in various files
The commit adds a new "Clock" class that enables time-based actions and execution. This change is reflected in Application.cpp, where the class is initialized. Additional changes were made to improve input state access, and more robust event tracking has been implemented in ImGuiLayer.cpp. DebugLayer now includes more actions for button presses. Other minor code cleanups and organizational changes have been made in various files.

+++ src/Ayla/Window/Window.h
@@ -5,7 +5,7 @@
 
 #include "Ayla/aypch.h"
 #include "Ayla/Events/Event.h"
-#include "GLFW/glfw3.h"
+
 
 using namespace Ayla::Events;
 
@@ -16,7 +16,7 @@ namespace Ayla::Windows{ // maybe change namespace to Display or Viewport instea
         float width;
         float height;
 
-        WindowProperties(std::string title = "Ayla Engine Window", int width = 1280, int height = 720) :
+        WindowProperties(std::string title = "Ayla Engine Window", float width = 1280.0f, float height = 720.0f) :
                 title(std::move(title)),
                 width(width),
                 height(height) {}
